mysite/requirements.txt
asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.1
netifaces==0.11.0
sqlparse==0.4.4
